+++ editor/src/core/frameworks/framework-hooks-default.ts
@@ -1,4 +1,4 @@
-import type { CreationDataFromProject } from '../model/project-import/project-import-utils'
+import type { CreationDataFromProject } from '../model/storyboard-utils'
 import type { FrameworkHooks } from './framework-hooks'
 
 export class DefaultFrameworkHooks implements FrameworkHooks {
